+++ packages/bluesky/lib/src/entities/embed_record_with_media.dart
@@ -19,10 +19,9 @@ class EmbedRecordWithMedia with _$EmbedRecordWithMedia {
   const factory EmbedRecordWithMedia({
     @Default('app.bsky.embed.recordWithMedia')
     @JsonKey(name: '\$type')
-        String type,
+    String type,
     required EmbedRecord record,
-    @EmbedMediaConverter()
-        required EmbedMedia media,
+    @EmbedMediaConverter() required EmbedMedia media,
   }) = _EmbedRecordWithMedia;
 
   factory EmbedRecordWithMedia.fromJson(Map<String, Object?> json) =>