+++ packages/bluesky/lib/src/entities/not_found_post.freezed.dart
@@ -3,7 +3,7 @@
 // ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
-part of 'post_thread_view_blocked.dart';
+part of 'not_found_post.dart';
 
 // **************************************************************************
 // FreezedGenerator
@@ -14,42 +14,40 @@ T _$identity<T>(T value) => value;
 final _privateConstructorUsedError = UnsupportedError(
     'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
-PostThreadViewBlocked _$PostThreadViewBlockedFromJson(
-    Map<String, dynamic> json) {
-  return _PostThreadViewBlocked.fromJson(json);
+NotFoundPost _$NotFoundPostFromJson(Map<String, dynamic> json) {
+  return _NotFoundPost.fromJson(json);
 }
 
 /// @nodoc
-mixin _$PostThreadViewBlocked {
+mixin _$NotFoundPost {
   @JsonKey(name: '\$type')
   String get type => throw _privateConstructorUsedError;
   @AtUriConverter()
   AtUri get uri => throw _privateConstructorUsedError;
-  bool get blocked => throw _privateConstructorUsedError;
+  bool get notFound => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
-  $PostThreadViewBlockedCopyWith<PostThreadViewBlocked> get copyWith =>
+  $NotFoundPostCopyWith<NotFoundPost> get copyWith =>
       throw _privateConstructorUsedError;
 }
 
 /// @nodoc
-abstract class $PostThreadViewBlockedCopyWith<$Res> {
-  factory $PostThreadViewBlockedCopyWith(PostThreadViewBlocked value,
-          $Res Function(PostThreadViewBlocked) then) =
-      _$PostThreadViewBlockedCopyWithImpl<$Res, PostThreadViewBlocked>;
+abstract class $NotFoundPostCopyWith<$Res> {
+  factory $NotFoundPostCopyWith(
+          NotFoundPost value, $Res Function(NotFoundPost) then) =
+      _$NotFoundPostCopyWithImpl<$Res, NotFoundPost>;
   @useResult
   $Res call(
       {@JsonKey(name: '\$type') String type,
       @AtUriConverter() AtUri uri,
-      bool blocked});
+      bool notFound});
 }
 
 /// @nodoc
-class _$PostThreadViewBlockedCopyWithImpl<$Res,
-        $Val extends PostThreadViewBlocked>
-    implements $PostThreadViewBlockedCopyWith<$Res> {
-  _$PostThreadViewBlockedCopyWithImpl(this._value, this._then);
+class _$NotFoundPostCopyWithImpl<$Res, $Val extends NotFoundPost>
+    implements $NotFoundPostCopyWith<$Res> {
+  _$NotFoundPostCopyWithImpl(this._value, this._then);
 
   // ignore: unused_field
   final $Val _value;
@@ -61,7 +59,7 @@ class _$PostThreadViewBlockedCopyWithImpl<$Res,
   $Res call({
     Object? type = null,
     Object? uri = null,
-    Object? blocked = null,
+    Object? notFound = null,
   }) {
     return _then(_value.copyWith(
       type: null == type
@@ -72,34 +70,34 @@ class _$PostThreadViewBlockedCopyWithImpl<$Res,
           ? _value.uri
           : uri // ignore: cast_nullable_to_non_nullable
               as AtUri,
-      blocked: null == blocked
-          ? _value.blocked
-          : blocked // ignore: cast_nullable_to_non_nullable
+      notFound: null == notFound
+          ? _value.notFound
+          : notFound // ignore: cast_nullable_to_non_nullable
               as bool,
     ) as $Val);
   }
 }
 
 /// @nodoc
-abstract class _$$_PostThreadViewBlockedCopyWith<$Res>
-    implements $PostThreadViewBlockedCopyWith<$Res> {
-  factory _$$_PostThreadViewBlockedCopyWith(_$_PostThreadViewBlocked value,
-          $Res Function(_$_PostThreadViewBlocked) then) =
-      __$$_PostThreadViewBlockedCopyWithImpl<$Res>;
+abstract class _$$_NotFoundPostCopyWith<$Res>
+    implements $NotFoundPostCopyWith<$Res> {
+  factory _$$_NotFoundPostCopyWith(
+          _$_NotFoundPost value, $Res Function(_$_NotFoundPost) then) =
+      __$$_NotFoundPostCopyWithImpl<$Res>;
   @override
   @useResult
   $Res call(
       {@JsonKey(name: '\$type') String type,
       @AtUriConverter() AtUri uri,
-      bool blocked});
+      bool notFound});
 }
 
 /// @nodoc
-class __$$_PostThreadViewBlockedCopyWithImpl<$Res>
-    extends _$PostThreadViewBlockedCopyWithImpl<$Res, _$_PostThreadViewBlocked>
-    implements _$$_PostThreadViewBlockedCopyWith<$Res> {
-  __$$_PostThreadViewBlockedCopyWithImpl(_$_PostThreadViewBlocked _value,
-      $Res Function(_$_PostThreadViewBlocked) _then)
+class __$$_NotFoundPostCopyWithImpl<$Res>
+    extends _$NotFoundPostCopyWithImpl<$Res, _$_NotFoundPost>
+    implements _$$_NotFoundPostCopyWith<$Res> {
+  __$$_NotFoundPostCopyWithImpl(
+      _$_NotFoundPost _value, $Res Function(_$_NotFoundPost) _then)
       : super(_value, _then);
 
   @pragma('vm:prefer-inline')
@@ -107,9 +105,9 @@ class __$$_PostThreadViewBlockedCopyWithImpl<$Res>
   $Res call({
     Object? type = null,
     Object? uri = null,
-    Object? blocked = null,
+    Object? notFound = null,
   }) {
-    return _then(_$_PostThreadViewBlocked(
+    return _then(_$_NotFoundPost(
       type: null == type
           ? _value.type
           : type // ignore: cast_nullable_to_non_nullable
@@ -118,9 +116,9 @@ class __$$_PostThreadViewBlockedCopyWithImpl<$Res>
           ? _value.uri
           : uri // ignore: cast_nullable_to_non_nullable
               as AtUri,
-      blocked: null == blocked
-          ? _value.blocked
-          : blocked // ignore: cast_nullable_to_non_nullable
+      notFound: null == notFound
+          ? _value.notFound
+          : notFound // ignore: cast_nullable_to_non_nullable
               as bool,
     ));
   }
@@ -128,14 +126,14 @@ class __$$_PostThreadViewBlockedCopyWithImpl<$Res>
 
 /// @nodoc
 @JsonSerializable()
-class _$_PostThreadViewBlocked implements _PostThreadViewBlocked {
-  const _$_PostThreadViewBlocked(
-      {@JsonKey(name: '\$type') required this.type,
+class _$_NotFoundPost implements _NotFoundPost {
+  const _$_NotFoundPost(
+      {@JsonKey(name: '\$type') this.type = 'app.bsky.feed.defs#notFoundPost',
       @AtUriConverter() required this.uri,
-      required this.blocked});
+      required this.notFound});
 
-  factory _$_PostThreadViewBlocked.fromJson(Map<String, dynamic> json) =>
-      _$$_PostThreadViewBlockedFromJson(json);
+  factory _$_NotFoundPost.fromJson(Map<String, dynamic> json) =>
+      _$$_NotFoundPostFromJson(json);
 
   @override
   @JsonKey(name: '\$type')
@@ -144,50 +142,50 @@ class _$_PostThreadViewBlocked implements _PostThreadViewBlocked {
   @AtUriConverter()
   final AtUri uri;
   @override
-  final bool blocked;
+  final bool notFound;
 
   @override
   String toString() {
-    return 'PostThreadViewBlocked(type: $type, uri: $uri, blocked: $blocked)';
+    return 'NotFoundPost(type: $type, uri: $uri, notFound: $notFound)';
   }
 
   @override
   bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _$_PostThreadViewBlocked &&
+            other is _$_NotFoundPost &&
             (identical(other.type, type) || other.type == type) &&
             (identical(other.uri, uri) || other.uri == uri) &&
-            (identical(other.blocked, blocked) || other.blocked == blocked));
+            (identical(other.notFound, notFound) ||
+                other.notFound == notFound));
   }
 
   @JsonKey(ignore: true)
   @override
-  int get hashCode => Object.hash(runtimeType, type, uri, blocked);
+  int get hashCode => Object.hash(runtimeType, type, uri, notFound);
 
   @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
-  _$$_PostThreadViewBlockedCopyWith<_$_PostThreadViewBlocked> get copyWith =>
-      __$$_PostThreadViewBlockedCopyWithImpl<_$_PostThreadViewBlocked>(
-          this, _$identity);
+  _$$_NotFoundPostCopyWith<_$_NotFoundPost> get copyWith =>
+      __$$_NotFoundPostCopyWithImpl<_$_NotFoundPost>(this, _$identity);
 
   @override
   Map<String, dynamic> toJson() {
-    return _$$_PostThreadViewBlockedToJson(
+    return _$$_NotFoundPostToJson(
       this,
     );
   }
 }
 
-abstract class _PostThreadViewBlocked implements PostThreadViewBlocked {
-  const factory _PostThreadViewBlocked(
-      {@JsonKey(name: '\$type') required final String type,
+abstract class _NotFoundPost implements NotFoundPost {
+  const factory _NotFoundPost(
+      {@JsonKey(name: '\$type') final String type,
       @AtUriConverter() required final AtUri uri,
-      required final bool blocked}) = _$_PostThreadViewBlocked;
+      required final bool notFound}) = _$_NotFoundPost;
 
-  factory _PostThreadViewBlocked.fromJson(Map<String, dynamic> json) =
-      _$_PostThreadViewBlocked.fromJson;
+  factory _NotFoundPost.fromJson(Map<String, dynamic> json) =
+      _$_NotFoundPost.fromJson;
 
   @override
   @JsonKey(name: '\$type')
@@ -196,9 +194,9 @@ abstract class _PostThreadViewBlocked implements PostThreadViewBlocked {
   @AtUriConverter()
   AtUri get uri;
   @override
-  bool get blocked;
+  bool get notFound;
   @override
   @JsonKey(ignore: true)
-  _$$_PostThreadViewBlockedCopyWith<_$_PostThreadViewBlocked> get copyWith =>
+  _$$_NotFoundPostCopyWith<_$_NotFoundPost> get copyWith =>
       throw _privateConstructorUsedError;
 }